refactor: move components to ui folder
