+++ frontend/svelte/src/lib/components/ui/Badge.svelte
