+++ frontend/svelte/src/lib/components/ui/InfiniteScroll.svelte
