+++ frontend/svelte/src/routes/Proposals.svelte
@@ -9,7 +9,7 @@
     listProposals,
   } from "../lib/utils/proposals.utils";
   import { proposalsStore } from "../lib/stores/proposals.store";
-  import InfiniteScroll from "../lib/components/InfiniteScroll.svelte";
+  import InfiniteScroll from "../lib/components/ui/InfiniteScroll.svelte";
   import ProposalCard from "../lib/components/ProposalCard.svelte";
   import Spinner from "../lib/components/ui/Spinner.svelte";
